feat: rename plugin
