+++ package/plugins/camera-lerp-plugin/camera-lerp-plugin.types.ts
@@ -2,6 +2,6 @@ import { MapControls } from "three/examples/jsm/controls/OrbitControls";
 
 import { IScenePlugin } from "../../types";
 
-export interface ITabNavigationTransitionPlugin extends IScenePlugin {
+export interface ICameraLerpPlugin extends IScenePlugin {
     controls: MapControls
 }
